// gateway/src/proto/tasks/generated/tasks_service.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.1
//   protoc               v3.19.1
// source: tasks_service.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "tasks";

export enum TaskStatus {
  TASK_STATUS_UNSPECIFIED = 0,
  TASK_STATUS_OPEN = 1,
  TASK_STATUS_IN_PROGRESS = 2,
  TASK_STATUS_DONE = 3,
  UNRECOGNIZED = -1,
}

export interface CreateTaskRequest {
  title: string;
  description: string;
  userId: string;
}

export interface CreateTaskResponse {
  task: TaskResponse | undefined;
}

export interface ListTasksRequest {
  status?: TaskStatus | undefined;
  searchQuery: string;
  userId: string;
}

export interface TaskResponse {
  id: string;
  title: string;
  description: string;
  status: TaskStatus;
}

export interface ListTasksResponse {
  tasks: TaskResponse[];
}

export interface GetTaskByIdRequest {
  id: string;
  userId: string;
}

export interface GetTaskByIdResponse {
  task: TaskResponse | undefined;
}

export interface UpdateTaskStatusRequest {
  id: string;
  status: TaskStatus;
  userId: string;
}

export interface UpdateTaskStatusResponse {
  task: TaskResponse | undefined;
}

export interface DeleteTaskRequest {
  id: string;
  userId: string;
}

export interface DeleteTaskResponse {
}

export const TASKS_PACKAGE_NAME = "tasks";

export interface TasksServiceClient {
  createTask(request: CreateTaskRequest): Observable<CreateTaskResponse>;

  listTasks(request: ListTasksRequest): Observable<ListTasksResponse>;

  getTaskById(request: GetTaskByIdRequest): Observable<GetTaskByIdResponse>;

  updateTaskStatus(request: UpdateTaskStatusRequest): Observable<UpdateTaskStatusResponse>;

  deleteTask(request: DeleteTaskRequest): Observable<DeleteTaskResponse>;
}

export interface TasksServiceController {
  createTask(
    request: CreateTaskRequest,
  ): Promise<CreateTaskResponse> | Observable<CreateTaskResponse> | CreateTaskResponse;

  listTasks(request: ListTasksRequest): Promise<ListTasksResponse> | Observable<ListTasksResponse> | ListTasksResponse;

  getTaskById(
    request: GetTaskByIdRequest,
  ): Promise<GetTaskByIdResponse> | Observable<GetTaskByIdResponse> | GetTaskByIdResponse;

  updateTaskStatus(
    request: UpdateTaskStatusRequest,
  ): Promise<UpdateTaskStatusResponse> | Observable<UpdateTaskStatusResponse> | UpdateTaskStatusResponse;

  deleteTask(
    request: DeleteTaskRequest,
  ): Promise<DeleteTaskResponse> | Observable<DeleteTaskResponse> | DeleteTaskResponse;
}

export function TasksServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createTask", "listTasks", "getTaskById", "updateTaskStatus", "deleteTask"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("TasksService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("TasksService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const TASKS_SERVICE_NAME = "TasksService";